Added static / dynamic light option
Note that if you're using a static light then modify its parameters,
you're shooting yourself in the foot and the behaviour is undefined.
Know what you're doing.
The static lights are pre-rendered only if the resulting texture can be
allocated. If it can you're safe, if it can't undefined behaviour will
occur.