Updated shadow algorithm
Implemented a new shadow algorithm that's about 50% on a single light. Multi light performance is severely degraded for now because of no exclusion logic.

Removed old ShadowSystem class as it is no longer needed.

The new algorithm introduces the concept of "filters" instead of walls. Future commits for more details